scramble to let tests compile
