+++ subs-aeagent/src/main/java/uk/ac/ebi/subs/arrayexpress/agent/ArrayExpressSubmissionProcessor.java
@@ -12,6 +12,7 @@ import uk.ac.ebi.subs.arrayexpress.model.ArrayExpressStudy;
 import uk.ac.ebi.subs.arrayexpress.model.SampleDataRelationship;
 import uk.ac.ebi.subs.arrayexpress.repo.ArrayExpressStudyRepository;
 import uk.ac.ebi.subs.arrayexpress.repo.SampleDataRelatioshipRepository;
+import uk.ac.ebi.subs.data.FullSubmission;
 import uk.ac.ebi.subs.data.Submission;
 import uk.ac.ebi.subs.data.submittable.Sample;
 import uk.ac.ebi.subs.processing.*;
@@ -112,7 +113,7 @@ public class ArrayExpressSubmissionProcessor {
 
     public List<ProcessingCertificate> processStudy(Study study, SubmissionEnvelope submissionEnvelope) {
         List<ProcessingCertificate> certs = new ArrayList<>();
-        Submission submission = submissionEnvelope.getSubmission();
+        FullSubmission submission = submissionEnvelope.getSubmission();
 
         if (!study.isAccessioned()) {
             study.setAccession("AE-MTAB-" + UUID.randomUUID());
@@ -153,7 +154,7 @@ public class ArrayExpressSubmissionProcessor {
     public List<ProcessingCertificate> processAssay(Assay assay, SubmissionEnvelope submissionEnvelope, ArrayExpressStudy arrayExpressStudy){
         List<ProcessingCertificate> certs = new ArrayList<>();
 
-        Submission submission = submissionEnvelope.getSubmission();
+        FullSubmission submission = submissionEnvelope.getSubmission();
 
         SampleDataRelationship sdr = new SampleDataRelationship();
         sdr.setId(UUID.randomUUID().toString());
