+++ subs-data-model/src/main/java/uk/ac/ebi/subs/data/submittable/Analysis.java
@@ -8,24 +8,12 @@ import java.util.List;
 
 public class Analysis extends AbstractSubsEntity<Analysis> implements Files {
 
-    @Id
-    String id;
-
     List<AnalysisRef> analysisRefs = new ArrayList<>();
     List<AssayDataRef> assayDataRefs = new ArrayList<>();
     List<StudyRef> studyRefs = new ArrayList<>();
     List<SampleRef> sampleRefs = new ArrayList<>();
     List<AssayRef> assayRefs = new ArrayList<>();
 
-    @Override
-    public String getId() {
-        return id;
-    }
-
-    @Override
-    public void setId(String id) {
-        this.id = id;
-    }
 
     public List<StudyRef> getStudyRefs() {
         return studyRefs;
