+++ subs-data-model/src/main/java/uk/ac/ebi/subs/data/submittable/Assay.java
@@ -9,22 +9,10 @@ import java.util.Objects;
 
 public class Assay extends AbstractSubsEntity<Assay> {
 
-    @Id
-    String id;
 
     StudyRef studyRef = new StudyRef();
     List<SampleUse> sampleUses = new ArrayList<>();
     List<ProtocolUse> protocolUses = new ArrayList<>();
-
-    @Override
-    public String getId() {
-        return id;
-    }
-
-    @Override
-    public void setId(String id) {
-        this.id = id;
-    }
 
     public StudyRef getStudyRef() {
         return studyRef;
