+++ subs-data-model/src/main/java/uk/ac/ebi/subs/data/submittable/EgaDataset.java
@@ -8,22 +8,11 @@ import java.util.List;
 
 public class EgaDataset extends AbstractSubsEntity<EgaDataset> {
 
-    @Id
-    String id;
+
 
     List<AssayDataRef> dataRefs = new ArrayList<>();
     List<AnalysisRef> analysisRefs = new ArrayList<>();
     EgaDacPolicyRef egaDacPolicyRef = new EgaDacPolicyRef();
-
-    @Override
-    public String getId() {
-        return id;
-    }
-
-    @Override
-    public void setId(String id) {
-        this.id = id;
-    }
 
     public List<AssayDataRef> getDataRefs() {
         return dataRefs;
