+++ subs-data-model/src/main/java/uk/ac/ebi/subs/data/submittable/Protocol.java
@@ -6,19 +6,6 @@ import uk.ac.ebi.subs.data.component.ProtocolRef;
 
 public class Protocol extends AbstractSubsEntity<Protocol> {
 
-    @Id
-    String id;
-
-    @Override
-    public String getId() {
-        return id;
-    }
-
-    @Override
-    public void setId(String id) {
-        this.id = id;
-    }
-
     @Override
     protected AbstractSubsRef<Protocol> newRef() {
         return new ProtocolRef();
