+++ subs-frontend/src/main/java/uk/ac/ebi/subs/frontend/config/RepositoryHandlerConfig.java
@@ -14,6 +14,7 @@ public class RepositoryHandlerConfig {
     }
 
     @Bean
-    SubmissionStudyEventHandler submissionStudyEventHandler() {return new SubmissionStudyEventHandler();}
+    SubmissionStudyEventHandler submissionStudyEventHandler() {
+        return new SubmissionStudyEventHandler();}
 
 }
