+++ subs-frontend/src/main/java/uk/ac/ebi/subs/frontend/handlers/SubmissionStudyEventHandler.java
@@ -4,10 +4,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.rest.core.annotation.HandleBeforeCreate;
-import org.springframework.data.rest.core.annotation.HandleBeforeSave;
 import org.springframework.data.rest.core.annotation.RepositoryEventHandler;
-import uk.ac.ebi.subs.repository.model.SubmissionStudy;
-import uk.ac.ebi.subs.repository.repo.SubmissionStudyRepo;
+import uk.ac.ebi.subs.data.submittable.Study;
+import uk.ac.ebi.subs.repository.submittable.StudyRepository;
 
 import java.util.UUID;
 
@@ -16,18 +15,18 @@ import java.util.UUID;
  * Repo event handler for studies nested in a submission
  *
  */
-@RepositoryEventHandler(SubmissionStudy.class)
+@RepositoryEventHandler(Study.class)
 public class SubmissionStudyEventHandler {
 
     private final Logger logger = LoggerFactory.getLogger(this.getClass());
 
     @Autowired
-    SubmissionStudyRepo submissionStudyRepo;
+    StudyRepository studyRepository;
 
     @HandleBeforeCreate
-    public void handleBeforeCreate(SubmissionStudy submissionStudy){
+    public void handleBeforeCreate(Study study){
         logger.warn("create");
-        submissionStudy.setId(UUID.randomUUID().toString());
+        study.setId(UUID.randomUUID().toString());
     }
