+++ subs-progress-monitor/src/test/java/util/Helpers.java
@@ -1,5 +1,6 @@
 package util;
 
+import uk.ac.ebi.subs.data.FullSubmission;
 import uk.ac.ebi.subs.data.component.Domain;
 import uk.ac.ebi.subs.data.submittable.Sample;
 import uk.ac.ebi.subs.data.Submission;
@@ -27,8 +28,8 @@ public class Helpers {
         return samples;
     }
 
-    public static Submission generateTestSubmission() {
-        Submission sub = new Submission();
+    public static FullSubmission generateTestSubmission() {
+        FullSubmission sub = new FullSubmission();
         sub.setId(UUID.randomUUID().toString());
         Domain d = new Domain();
         d.setName("subs-test");
