+++ subs-repository/src/main/java/uk/ac/ebi/subs/repository/SubmissionRepository.java
@@ -6,10 +6,15 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.data.mongodb.repository.Query;
 import org.springframework.data.repository.query.Param;
+import org.springframework.data.rest.core.annotation.Description;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import org.springframework.data.rest.core.annotation.RestResource;
 import uk.ac.ebi.subs.data.Submission;
 
+@RepositoryRestResource(
+        collectionResourceRel = "submissions",
+        path = "submissions",
+        collectionResourceDescription = @Description("Metadata submissions to the EBI archives"))
 public interface SubmissionRepository extends MongoRepository<Submission,String>{
 
     @Query(value="{ 'domain.name' : ?0 }")
